+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated pipeline Windows VM image to windows-2019
+
 ## [0.4.0] - 2020-03-09
 
 ### Added