+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the cmdlet `Assert-BoundParameter`. _This cndlet comes from_
   _ComputerManagementDsc._
+- Added GitHub templates in the repository to help contributors.
 
 ### Changed