+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2020-04-23
+
 ### Added
 
 - Added the cmdlet `Assert-BoundParameter`. _This cmdlet comes from_